+++ src/main/kotlin/vendor1/processor/BuyDrinkCommandProcessor.kt
@@ -1,17 +1,18 @@
-package vendor1.command
+package vendor1.processor
 
 import java.io.OutputStream
 
-class PrintSpecificationCommandProcessor : CommandProcessor {
+class BuyDrinkCommandProcessor : CommandProcessor {
 
     override suspend fun sendResponse(command: String, outputStream: OutputStream): Boolean {
         try {
-            val result = vendorService.printSpecification(command) ?: return false
+            val result = vendorService.buyDrink(command) ?: return false
             outputStream.write(toResponseData(result))
             outputStream.flush()
             return true
         } catch (e: Exception) {
-            throw RuntimeException(e.cause)
+            throw RuntimeException(e)
         }
+
     }
 }
