+++ src/main/kotlin/vendor1/processor/PrintSpecificationCommandProcessor.kt
@@ -1,12 +1,12 @@
-package vendor1.command
+package vendor1.processor
 
 import java.io.OutputStream
 
-class RegisterCommandProcessor : CommandProcessor {
+class PrintSpecificationCommandProcessor : CommandProcessor {
 
     override suspend fun sendResponse(command: String, outputStream: OutputStream): Boolean {
         try {
-            val result = vendorService.registerDrink(command) ?: return false
+            val result = vendorService.printSpecification(command) ?: return false
             outputStream.write(toResponseData(result))
             outputStream.flush()
             return true
