+++ src/main/kotlin/vendor1/processor/RegisterCommandProcessor.kt
@@ -1,18 +1,17 @@
-package vendor1.command
+package vendor1.processor
 
 import java.io.OutputStream
 
-class BuyDrinkCommandProcessor : CommandProcessor {
+class RegisterCommandProcessor : CommandProcessor {
 
     override suspend fun sendResponse(command: String, outputStream: OutputStream): Boolean {
         try {
-            val result = vendorService.buyDrink(command) ?: return false
+            val result = vendorService.registerDrink(command) ?: return false
             outputStream.write(toResponseData(result))
             outputStream.flush()
             return true
         } catch (e: Exception) {
-            throw RuntimeException(e)
+            throw RuntimeException(e.cause)
         }
-
     }
 }
