+++ src/main/kotlin/vendor1/processor/StatusCommandProcessor.kt
@@ -1,4 +1,4 @@
-package vendor1.command
+package vendor1.processor
 
 import java.io.OutputStream
